+++ minder_utils/feature_engineering/adding_features.py
@@ -1,40 +1,6 @@
-from scipy.stats import entropy as cal_entropy
 from minder_utils.configurations import feature_config, config
 from .util import *
 import pandas as pd
-
-
-def calculate_entropy(df: pd.DataFrame, sensors: list) -> pd.DataFrame:
-    '''
-    Return a dataframe with research id, week id, and entropy value
-    based on list of sensors given. If resulting activity count of
-    any of the sensors given in the list is zero, the value of the
-    entropy will be NaN.
-
-    Args:
-        df: Dataframe, contains at least 4 columns ['id', 'week', 'location', 'value']
-        sensors: List, the
-
-    Returns:
-
-    '''
-    # Filter the sensors
-    df = df[df.location.isin(sensors)]
-    # Sum the the number of readings of sensors weekly
-    sensor_summation = df.groupby(['id', 'week'])['value'].sum().reset_index()
-    sensor_summation.columns = ['id', 'week', 'summation']
-
-    # Merge with existing dataframe
-    df = pd.merge(df, sensor_summation)
-
-    # Calculate the probabilities
-    df['probabilities'] = df['value'] / df['summation']
-
-    # Calculate the entropy
-    df = df.groupby(['id', 'week'])['probabilities'].apply(cal_entropy).reset_index()
-    df.columns = ['id', 'week', 'value']
-    df['location'] = 'entropy'
-    return df
 
 
 def get_bathroom_activity(data, time_range, name):
