+++ minder_utils/feature_engineering/calculation.py
@@ -1,7 +1,9 @@
 from scipy.stats import ks_2samp
+import pandas as pd
+from scipy.stats import entropy as cal_entropy
 
 
-def weekly_compare(df, func, num_previous_week=1):
+def weekly_compare(df: pd.DataFrame, func, num_previous_week=1) -> dict:
     '''
     Function to compare the values of each patient in current week to previous weeks
     Args:
@@ -37,7 +39,7 @@ def weekly_compare(df, func, num_previous_week=1):
     return results
 
 
-def threshold_compare(df, func='>', threshold=36):
+def threshold_compare(df: pd.DataFrame, func='>', threshold=36) -> pd.DataFrame:
     '''
     Function to filter the dataframe by threshold
     Args:
@@ -54,7 +56,39 @@ def threshold_compare(df, func='>', threshold=36):
         return df[df.value < threshold]
 
 
+def calculate_entropy(df: pd.DataFrame, sensors: list) -> pd.DataFrame:
+    '''
+    Return a dataframe with research id, week id, and entropy value
+    based on list of sensors given. If resulting activity count of
+    any of the sensors given in the list is zero, the value of the
+    entropy will be NaN.
+
+    Args:
+        df: Dataframe, contains at least 4 columns ['id', 'week', 'location', 'value']
+        sensors: List, the
+
+    Returns:
+
+    '''
+    assert len(sensors) >= 2, 'need at least two sensors to calculate the entropy'
+    # Filter the sensors
+    df = df[df.location.isin(sensors)]
+    # Sum the the number of readings of sensors weekly
+    sensor_summation = df.groupby(['id', 'week'])['value'].sum().reset_index()
+    sensor_summation.columns = ['id', 'week', 'summation']
+
+    # Merge with existing dataframe
+    df = pd.merge(df, sensor_summation)
+
+    # Calculate the probabilities
+    df['probabilities'] = df['value'] / df['summation']
+
+    # Calculate the entropy
+    df = df.groupby(['id', 'week'])['probabilities'].apply(cal_entropy).reset_index()
+    df.columns = ['id', 'week', 'value']
+    df['location'] = 'entropy'
+    return df
+
+
 def kolmogorov_smirnov(freq1, freq2):
     return ks_2samp(freq1, freq2)
-
-
